Implement creating/deleting addresses.

+++ api/src/routes/merchants.addresses.routes.js
@@ -5,6 +5,7 @@ import { Router } from 'express';
 import { Merchant, Address } from '../database/models';
 import checkJwt from '../middleware/authentication.js';
 import adminRole from '../middleware/admin.auth.js';
+import { Utils } from '../util.js';
 
 const router = Router({mergeParams: true});
 
@@ -12,7 +13,7 @@ const router = Router({mergeParams: true});
 router.post("/", checkJwt, adminRole, async (req, res) => {
     try {
         const merchantId = req.params.merchantId;
-        const { address, city, province, country, postalcode } = req.body;
+        const { address, city, province, country, postalcode, neighbourhood } = req.body;
 
         if (!address || !city || !province || !country || !postalcode) {
             return res.status(400).json({ message: 'address, province, country, and postalcode are required if specifying an address' });
@@ -26,6 +27,20 @@ router.post("/", checkJwt, adminRole, async (req, res) => {
         const merchant = await Merchant.findOne({
             where: {id : merchantId }
         });
+        if (!merchant) {
+            return res.status(404).json({ message: 'The merchant with the given id was not found.'});
+        }
+
+        let addrObj = {
+            address1,
+            address2,
+            address3,
+            city,
+            province,
+            country,
+            postalcode,
+            neighbourhood
+        }
         const newAddress = await Address.create({
             address1,
             address2,
@@ -33,10 +48,12 @@ router.post("/", checkJwt, adminRole, async (req, res) => {
             city,
             province,
             country,
-            postalcode
+            postalcode,
+            neighbourhood,
+            full: Utils.getFullAddress(addrObj)
         });
         merchant.addAddress(newAddress).then(() => {
-            return res.status(201).json({ merchant, newAddress });
+            return res.status(201).json({ merchant, address: newAddress });
         }).catch((error) => {
             return res.status(500).json({ message: error });
         });
@@ -98,11 +115,15 @@ router.put("/:id", checkJwt, adminRole, async (req, res) => {
 // Delete an address by id
 router.delete("/:id", checkJwt, adminRole, async (req, res) => {
     try {
-        const address = await Address.destroy({ where: { id: req.params.id } });
-        if (!address)
-            return res.status(404).json({ message: 'The address with the given id was not found' });
-    
-        return res.status(200).json({ message: 'The address was deleted successfully' });
+        const merchantId = req.params.merchantId;
+        const merchant = await Merchant.findOne({ where: {id: merchantId}, include: {model: Address, where: {id: req.params.id}}});
+        if (merchant && merchant.Addresses && Array.isArray(merchant.Addresses)) {
+            const address = await Address.destroy({ where: { id: req.params.id } });
+            if (!address)
+                return res.status(404).json({ message: 'The address with the given id was not found' });
+            return res.status(200).json({ message: 'The address was deleted successfully' });
+        }
+        return res.status(404).json({ message: 'The address with the given id is not associated with the given merchant.'});
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
